blobs: Close file before calling 'AgentSynchronizeData'
The IPMI BLOB handler fills 'pcie' file with content and calls
'AgentSynchronizeData' method for synchronization.
Because currently the file is not closed before the call to
'AgentSynchronizeData', this can lead to errors that smbios-mdr tries
to parse file that was not yet flushed to the disk.
Add file close operation to fix the issue.

Signed-off-by: Konstantin Aladyshev <aladyshev22@gmail.com>